+++ src/index.ts
@@ -1,3 +1,4 @@
+import mongoose from 'mongoose';
 import { DecoratorKeys } from './internal/constants';
 import { getName } from './internal/utils';
 import { logger } from './logSettings';
@@ -13,13 +14,10 @@ import type { BeAnObject, IIndexArray, IndexOptions } from './types';
  * class ClassName {}
  * ```
  */
-export function index<T extends BeAnObject = BeAnObject>(
-  fields: Partial<Record<keyof T, string | -1 | 1>>,
-  options?: IndexOptions<T>
-): ClassDecorator {
+export function index<T extends BeAnObject = BeAnObject>(fields: mongoose.IndexDefinition, options?: IndexOptions<T>): ClassDecorator {
   return (target: any) => {
     logger.info('Adding "%o" Indexes to %s', { fields, options }, getName(target));
-    const indices: IIndexArray<any>[] = Array.from(Reflect.getMetadata(DecoratorKeys.Index, target) ?? []);
+    const indices: IIndexArray[] = Array.from(Reflect.getMetadata(DecoratorKeys.Index, target) ?? []);
     indices.push({ fields, options });
     Reflect.defineMetadata(DecoratorKeys.Index, indices, target);
   };