+++ src/types.ts
@@ -12,9 +12,18 @@ import type { Severity, PropType } from './internal/constants';
  * ```
  */
 export type DocumentType<T, QueryHelpers = BeAnObject> = (T extends { _id: unknown }
-  ? mongoose.Document<T['_id'], QueryHelpers> & T
-  : mongoose.Document<any, QueryHelpers> & T) &
+  ? mongoose.Document<T['_id'], QueryHelpers, T>
+  : mongoose.Document<any, QueryHelpers, T>) &
+  T &
   IObjectWithTypegooseFunction;
+/**
+ * Get the Type of an instance of a SubDocument with Class properties
+ */
+export type SubDocumentType<T, QueryHelpers = BeAnObject> = DocumentType<T, QueryHelpers> & mongoose.Types.Subdocument;
+/**
+ * Get the Type of an instance of a SubDocument that exists within an array, with Class properties
+ */
+export type ArraySubDocumentType<T, QueryHelpers = BeAnObject> = DocumentType<T, QueryHelpers> & mongoose.Types.ArraySubdocument;
 /**
  * Used Internally for ModelTypes
  */
@@ -24,7 +33,9 @@ export type ModelType<T, QueryHelpers = BeAnObject> = mongoose.Model<DocumentTyp
  */
 export type AnyParamConstructor<T> = new (...args: any) => T;
 /**
- * The Type of a Model that gets returned by "getModelForClass" and "setModelForClass"
+ * The Type for Models used in typegoose, mostly returned by "getModelForClass" and "addModelToTypegoose"
+ * @example
+ * const Model: ReturnModelType<typeof YourClass, YourClassQueryHelper> = mongoose.model("YourClass", YourClassSchema);
  */
 export type ReturnModelType<U extends AnyParamConstructor<any>, QueryHelpers = BeAnObject> = ModelType<InstanceType<U>, QueryHelpers> & U;
 
@@ -62,12 +73,18 @@ export interface BasePropOptions {
    */
   required?: mongoose.SchemaTypeOptions<any>['required'];
   /** Only accept Values from the Enum(|Array) */
-  enum?: string[] | BeAnObject;
+  enum?: mongoose.SchemaTypeOptions<any>['enum'];
   /**
    * Add "null" to the enum array
    * Note: Custom Typegoose Option
    */
   addNullToEnum?: boolean;
+  /**
+   * Set Custom "warnMixed" Severity for a specific property
+   * Overwrites Severity set in "modelOptions" for a specific property
+   * Note: Custom Typegoose Option
+   */
+  allowMixed?: Severity;
   /** Give the Property a default Value */
   default?: mongoose.SchemaTypeOptions<any>['default']; // i know this one does not have much of an effect, because of "any"
   /** Give an Validator RegExp or Function */
@@ -331,7 +348,7 @@ export interface VirtualOptions {
   /** Extra Query Options */
   options?: mongoose.VirtualTypeOptions['options'];
   /** Match Options */
-  match?: KeyStringAny | ((doc) => KeyStringAny);
+  match?: mongoose.VirtualTypeOptions['match'];
   /**
    * If you set this to `true`, Mongoose will call any custom getters you defined on this virtual.
    *
@@ -381,7 +398,7 @@ export type Ref<
 > = mongoose.PopulatedDoc<PopulatedType, RawId>;
 
 /**
- * An Function type for a function that doesn't have any arguments and doesn't return anything
+ * A Function type for a function that doesn't have any arguments and doesn't return anything
  */
 export type EmptyVoidFn = () => void;
 
@@ -450,9 +467,7 @@ export type DecoratedPropertyMetadataMap = Map<string | symbol, DecoratedPropert
 /**
  * copy-paste from mongodb package (should be same as IndexOptions from 'mongodb')
  */
-export interface IndexOptions<T> extends mongoose.IndexOptions {
-  weights?: Partial<Record<keyof T, number>>;
-}
+export type IndexOptions<_T> = mongoose.IndexOptions; // TODO: remove unused generic in typegoose 10
 
 /**
  * Type for the Values stored in the Reflection for Indexes
@@ -461,21 +476,21 @@ export interface IndexOptions<T> extends mongoose.IndexOptions {
  * const indices: IIndexArray[] = Reflect.getMetadata(DecoratorKeys.Index, target) || []);
  * ```
  */
-export interface IIndexArray<T> {
+export interface IIndexArray {
   fields: KeyStringAny;
-  options?: IndexOptions<T>;
+  options?: IndexOptions<unknown>;
 }
 
 /**
  * Type for the Values stored in the Reflection for Plugins
  * @example
  * ```ts
- * const plugins: IPluginsArray<any>[] = Array.from(Reflect.getMetadata(DecoratorKeys.Plugins, target) ?? []);
+ * const plugins: IPluginsArray[] = Array.from(Reflect.getMetadata(DecoratorKeys.Plugins, target) ?? []);
  * ```
  */
-export interface IPluginsArray<T> {
+export interface IPluginsArray {
   mongoosePlugin: Func;
-  options: T;
+  options: BeAnObject | undefined;
 }
 
 /**
@@ -485,7 +500,7 @@ export interface IPluginsArray<T> {
  * const virtuals: VirtualPopulateMap = new Map(Reflect.getMetadata(DecoratorKeys.VirtualPopulate, target.constructor) ?? []);
  * ```
  */
-export type VirtualPopulateMap = Map<string, any & VirtualOptions>;
+export type VirtualPopulateMap = Map<string, VirtualOptions & Record<string, unknown>>;
 
 /**
  * Gets the signature (parameters with their types, and the return type) of a function type.
@@ -577,7 +592,10 @@ export interface IObjectWithTypegooseFunction {
   typegooseName(): string;
 }
 
-/** Interface describing a Object that has a "typegooseName" Value */
+/**
+ * Interface describing a Object that has a "typegooseName" Value
+ * @deprecated This interface and value will be removed in typegoose 10, use {@link IObjectWithTypegooseFunction} instead
+ */
 export interface IObjectWithTypegooseName {
   typegooseName: string;
 }
@@ -587,10 +605,8 @@ export interface IPrototype {
   prototype?: any;
 }
 
-/** An Helper Interface for key: any: string */
-export interface KeyStringAny {
-  [key: string]: any;
-}
+/** An Helper Interface for defining a "key" index of "string" and "value" of "any" */
+export type KeyStringAny = Record<string, any>;
 
 /**
  * The Return Type of "utils.getType"
@@ -601,6 +617,7 @@ export interface GetTypeReturn {
 }
 
 /**
- * This type is for lint error "ban-types"
+ * This type is for lint error "ban-types" where "{}" would be used
+ * This type is seperate from "{@link KeyStringAny}" because it has a different meaning
  */
 export type BeAnObject = Record<string, any>;