+++ website/sidebars.js
@@ -6,7 +6,7 @@ module.exports = {
       'api/decorators/model-options',
       'api/decorators/hooks',
       'api/decorators/indexes',
-      'api/decorators/plugins',
+      'api/decorators/plugin',
       'api/decorators/query-method',
     ],
     Functions: [
@@ -54,6 +54,7 @@ module.exports = {
       'guides/advanced/non-nested-discriminators',
       'guides/advanced/nested-discriminators',
       'guides/advanced/name-generation',
+      'guides/advanced/manual-schema-modification',
     ],
     Migration: ['guides/migration/migrate-9', 'guides/migration/migrate-8', 'guides/migration/migrate-7', 'guides/migration/migrate-6'],
   },